third_party/blink/web_tests/fast/domurl/check-instanceof-domurl-functions-expected.txt

Test instanceof functions and properties of URL and webkitURL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'createObjectURL' in URL is true
PASS 'revokeObjectURL' in URL is true
PASS url.createObjectURL is undefined.
PASS url.revokeObjectURL is undefined.
PASS window.URL.createObjectURL is defined.
PASS window.URL.revokeObjectURL is defined.
PASS 'createObjectURL' in webkitURL is true
PASS 'revokeObjectURL' in webkitURL is true
PASS window.webkitURL.createObjectURL is defined.
PASS window.webkitURL.revokeObjectURL is defined.
PASS successfullyParsed is true

TEST COMPLETE
